Bugs: fixed, tests needed sorely
